Mover clases al core y enum a la carpeta Enums
