+++ src/dominio/Core/Obtener_Zapato.java
@@ -1,7 +1,9 @@
-package dominio;
+package dominio.Core;
 
 import java.util.ArrayList;
 import java.util.stream.Collectors;
+
+import dominio.GestorZapatos;
 
 public class Obtener_Zapato {
