+++ src/dominio/Core/Obtener_Zapatos_hasta_talle.java
@@ -1,8 +1,10 @@
-package dominio;
+package dominio.Core;
 
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.stream.Collectors;
+
+import dominio.GestorZapatos;
 
 public class Obtener_Zapatos_hasta_talle {
