+++ src/dominio/Core/Precio.java
@@ -1,4 +1,4 @@
-package dominio;
+package dominio.Core;
 
 import java.text.SimpleDateFormat;
 import java.util.Calendar;
