+++ src/dominio/Enums/Color.java
@@ -1,4 +1,4 @@
-package dominio;
+package dominio.Enums;
 
 public enum Color {
 	ROJO,NEGRO,AZUL,BLANCO;
