+++ src/dominio/Enums/Pais.java
@@ -1,4 +1,4 @@
-package dominio;
+package dominio.Enums;
 
 public enum Pais {
     ITALIA,ARGENTINA,MACEDONIA,URUGUAY;
